refactor(engine): move handle key logic from player
