+++ packages/client/src/core/engine.ts
@@ -16,14 +16,7 @@ import heroImage from '../assets/hero_run.png'
 import { calcPosition, randomFromInterval } from './utils'
 import { EnemyType, KeyConfiguration } from './types'
 
-// FIXME этому тут не место
-function toggleFullScreen() {
-  if (!document.fullscreenElement) {
-    document.documentElement.requestFullscreen();
-  } else if (document.exitFullscreen) {
-    document.exitFullscreen();
-  }
-}
+
 
 export class Engine {
   get gameOver(): boolean {
@@ -45,7 +38,6 @@ export class Engine {
   private score = 0
   private enemies: Enemy[] = []
   private enemyTimer = 0
-  private keyConfig: KeyConfiguration = { Space: 'up', Enter: 'up' }
 
   private readonly _isGameStartWords = ['3...', '2...', '1...', 'Go']
   private readonly _isGameStartDelayWord = 1000
@@ -83,7 +75,7 @@ export class Engine {
     })
   }
 
-  game = (ctx: CanvasRenderingContext2D, deltaTime: number) => {
+  game = (ctx: CanvasRenderingContext2D, deltaTime: number, keyConfig: KeyConfiguration) => {
     ctx.clearRect(0, 0, ctx.canvas.width, ctx.canvas.height)
 
     this.background.draw(ctx)
@@ -92,7 +84,7 @@ export class Engine {
     this.checkCollisions()
 
     this.player.draw(ctx)
-    this.player.update(this.keyConfig, deltaTime)
+    this.player.update(keyConfig, deltaTime)
 
     this.displayScore(ctx)
 
@@ -108,26 +100,6 @@ export class Engine {
 
     if (this.gameOver) {
       this.displayGameOver(ctx)
-    }
-  }
-
-  handleKeyDown = (e: Event) => {
-    const { code } = e as KeyboardEvent
-    if (code === 'Space') {
-      this.keyConfig.Space = 'down'
-    } else if (code === 'Enter') {
-      this.keyConfig.Enter = 'down'
-      toggleFullScreen()
-    }
-  }
-
-  handleKeyUp = (e: Event) => {
-    const { code } = e as KeyboardEvent
-    if (code === 'Space') {
-      this.keyConfig.Space = 'up'
-    } else if (code === 'Enter') {
-      this.keyConfig.Enter = 'up'
-      this.keyConfig.Enter = 'up'
     }
   }
